+++ src/api/controllers/instance.controller.ts
@@ -1,4 +1,4 @@
-import { delay } from '@whiskeysockets/baileys';
+import { delay } from 'baileys';
 import { isURL } from 'class-validator';
 import EventEmitter2 from 'eventemitter2';
 import { v4 } from 'uuid';
@@ -15,12 +15,12 @@ import { WebsocketService } from '../integrations/websocket/services/websocket.s
 import { RepositoryBroker } from '../repository/repository.manager';
 import { AuthService, OldToken } from '../services/auth.service';
 import { CacheService } from '../services/cache.service';
+import { BaileysStartupService } from '../services/channels/whatsapp.baileys.service';
+import { BusinessStartupService } from '../services/channels/whatsapp.business.service';
 import { IntegrationService } from '../services/integration.service';
 import { WAMonitoringService } from '../services/monitor.service';
 import { SettingsService } from '../services/settings.service';
 import { WebhookService } from '../services/webhook.service';
-import { BaileysStartupService } from '../services/whatsapp/whatsapp.baileys.service';
-import { BusinessStartupService } from '../services/whatsapp/whatsapp.business.service';
 import { Events, Integration, wa } from '../types/wa.types';
 import { ProxyController } from './proxy.controller';
 
@@ -65,6 +65,7 @@ export class InstanceController {
     chatwoot_reopen_conversation,
     chatwoot_conversation_pending,
     chatwoot_import_contacts,
+    chatwoot_name_inbox,
     chatwoot_import_messages,
     chatwoot_days_limit_import_messages,
     reject_call,
@@ -513,7 +514,7 @@ export class InstanceController {
           token: chatwoot_token,
           url: chatwoot_url,
           sign_msg: chatwoot_sign_msg || false,
-          name_inbox: instance.instanceName.split('-cwId-')[0],
+          name_inbox: chatwoot_name_inbox ?? instance.instanceName.split('-cwId-')[0],
           number,
           reopen_conversation: chatwoot_reopen_conversation || false,
           conversation_pending: chatwoot_conversation_pending || false,
@@ -577,7 +578,7 @@ export class InstanceController {
           import_messages: chatwoot_import_messages ?? true,
           days_limit_import_messages: chatwoot_days_limit_import_messages || 60,
           number,
-          name_inbox: instance.instanceName,
+          name_inbox: chatwoot_name_inbox ?? instance.instanceName,
           webhook_url: `${urlServer}/chatwoot/webhook/${encodeURIComponent(instance.instanceName)}`,
         },
       };