+++ src/api/services/channel.service.ts
@@ -1,5 +1,5 @@
-import { WASocket } from '@whiskeysockets/baileys';
 import axios from 'axios';
+import { WASocket } from 'baileys';
 import { execSync } from 'child_process';
 import { isURL } from 'class-validator';
 import EventEmitter2 from 'eventemitter2';
@@ -38,17 +38,17 @@ import { waMonitor } from '../server.module';
 import { Events, wa } from '../types/wa.types';
 import { CacheService } from './cache.service';
 
-export class WAStartupService {
+export class ChannelStartupService {
   constructor(
     public readonly configService: ConfigService,
     public readonly eventEmitter: EventEmitter2,
     public readonly repository: RepositoryBroker,
     public readonly chatwootCache: CacheService,
   ) {
-    this.logger.verbose('WAStartupService initialized');
+    this.logger.verbose('ChannelStartupService initialized');
   }
 
-  public readonly logger = new Logger(WAStartupService.name);
+  public readonly logger = new Logger(ChannelStartupService.name);
 
   public client: WASocket;
   public readonly instance: wa.Instance = {};
@@ -136,11 +136,13 @@ export class WAStartupService {
 
   public async findIntegration() {
     this.logger.verbose('Finding integration');
-    const data = await this.repository.integration.find(this.instanceName);
+    let data: any;
+
+    data = await this.repository.integration.find(this.instanceName);
 
     if (!data) {
-      this.logger.verbose('Integration not found');
-      throw new NotFoundException('Integration not found');
+      this.repository.integration.create({ integration: 'WHATSAPP-BAILEYS', number: '', token: '' }, this.instanceName);
+      data = { integration: 'WHATSAPP-BAILEYS', number: '', token: '' };
     }
 
     this.logger.verbose(`Integration: ${data.integration}`);
@@ -740,7 +742,7 @@ export class WAStartupService {
 
           if (this.configService.get<Log>('LOG').LEVEL.includes('WEBHOOKS')) {
             const logData = {
-              local: WAStartupService.name + '.sendData-RabbitMQ',
+              local: ChannelStartupService.name + '.sendData-RabbitMQ',
               event,
               instance: this.instance.name,
               data,
@@ -796,7 +798,7 @@ export class WAStartupService {
           sqs.sendMessage(params, (err, data) => {
             if (err) {
               this.logger.error({
-                local: WAStartupService.name + '.sendData-SQS',
+                local: ChannelStartupService.name + '.sendData-SQS',
                 message: err?.message,
                 hostName: err?.hostname,
                 code: err?.code,
@@ -808,7 +810,7 @@ export class WAStartupService {
             } else {
               if (this.configService.get<Log>('LOG').LEVEL.includes('WEBHOOKS')) {
                 const logData = {
-                  local: WAStartupService.name + '.sendData-SQS',
+                  local: ChannelStartupService.name + '.sendData-SQS',
                   event,
                   instance: this.instance.name,
                   data,
@@ -852,7 +854,7 @@ export class WAStartupService {
 
         if (this.configService.get<Log>('LOG').LEVEL.includes('WEBHOOKS')) {
           const logData = {
-            local: WAStartupService.name + '.sendData-WebsocketGlobal',
+            local: ChannelStartupService.name + '.sendData-WebsocketGlobal',
             event,
             instance: this.instance.name,
             data,
@@ -882,7 +884,7 @@ export class WAStartupService {
 
         if (this.configService.get<Log>('LOG').LEVEL.includes('WEBHOOKS')) {
           const logData = {
-            local: WAStartupService.name + '.sendData-Websocket',
+            local: ChannelStartupService.name + '.sendData-Websocket',
             event,
             instance: this.instance.name,
             data,
@@ -916,7 +918,7 @@ export class WAStartupService {
 
         if (this.configService.get<Log>('LOG').LEVEL.includes('WEBHOOKS')) {
           const logData = {
-            local: WAStartupService.name + '.sendDataWebhook-local',
+            local: ChannelStartupService.name + '.sendDataWebhook-local',
             url: baseURL,
             event,
             instance: this.instance.name,
@@ -956,7 +958,7 @@ export class WAStartupService {
           }
         } catch (error) {
           this.logger.error({
-            local: WAStartupService.name + '.sendDataWebhook-local',
+            local: ChannelStartupService.name + '.sendDataWebhook-local',
             message: error?.message,
             hostName: error?.hostname,
             syscall: error?.syscall,
@@ -988,7 +990,7 @@ export class WAStartupService {
 
         if (this.configService.get<Log>('LOG').LEVEL.includes('WEBHOOKS')) {
           const logData = {
-            local: WAStartupService.name + '.sendDataWebhook-global',
+            local: ChannelStartupService.name + '.sendDataWebhook-global',
             url: globalURL,
             event,
             instance: this.instance.name,
@@ -1027,7 +1029,7 @@ export class WAStartupService {
           }
         } catch (error) {
           this.logger.error({
-            local: WAStartupService.name + '.sendDataWebhook-global',
+            local: ChannelStartupService.name + '.sendDataWebhook-global',
             message: error?.message,
             hostName: error?.hostname,
             syscall: error?.syscall,