+++ src/utils/use-multi-file-auth-state-redis-db.ts
@@ -1,10 +1,4 @@
-import {
-  AuthenticationCreds,
-  AuthenticationState,
-  initAuthCreds,
-  proto,
-  SignalDataTypeMap,
-} from '@whiskeysockets/baileys';
+import { AuthenticationCreds, AuthenticationState, initAuthCreds, proto, SignalDataTypeMap } from 'baileys';
 
 import { CacheService } from '../api/services/cache.service';
 import { Logger } from '../config/logger.config';