+++ CHANGELOG.md
@@ -1,9 +1,12 @@
-# 1.7.4 (develop)
+# 1.7.4 (2024-04-28 09:46)
 
 ### Fixed
 * Adjusts in proxy on fetchAgent
 * Recovering messages lost with redis cache
 * Log when init redis cache service
+* Recovering messages lost with redis cache
+* Chatwoot inbox name
+* Update Baileys version
 
 # 1.7.3 (2024-04-18 12:07)